requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
croniter==6.0.0
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.121.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-openai==1.0.2
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langsmith==0.4.41
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
openai==2.7.1
orjson==3.11.4
packaging==25.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.0
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pypdf==6.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
redis==7.0.1
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rq==2.6.0
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
yarl==1.22.0
zstandard==0.25.0
